requirements.txt
streamlit==1.28.0
librosa==0.10.1
numpy==1.26.4
tensorflow-cpu==2.16.1
scikit-learn==1.4.2
matplotlib==3.9.0
resampy==0.4.2
soundfile==0.12.1